First attempt at getting a render component working.
example/src/component/mesh-renderermesh-renderer.lisp is the
component.

It currently does some bad shit at toplevel, like defvao and whatnot.
That's all temrporary code.

Added a shader for the experimental mesh-renderer which expects the
(pos color uv) interleaved layout used in the defvao.

Had to add a reference to the camera and shaders in the context-table so
the mesh-renderer render-component protocol method would have access to them.
That might be able to change too in the future, once the rendering works and
we can start to see the pattern.