Checkpoint commit for mesh dsl.
This is ugly -- really ugly. It definitely needs a once-over, but the basic
functionality is now in place. We are able to create a glkit VAO declaration
given our additive layout groups merged into a vertex spec. Divisors still are
not handled -- currently everything is treated as divisor 0. That will be the
next step, along with some major cleanups, before I move onto the API for
writing buffer data given a buffer name, rather than an integral index.