Reasonably completed the core mesh component to cache meshes that it finds at particular locations. Now, it only makes new vaos when it encounters new meshes it hadn't seen before. The comments and code in mesh's INITIALIZE-COMPONENT should explain how it works.
Added a new file which contains a couple of CLOS objects that aare used
for the shared-storage for the mesh-type.

Ensured that each mesh loaded by the mesh component has its own
copy of the vao-id and the layout associated with the mesh at that
location. This allows the caching of meshes with different layouts. (So
as long as the mesh-renderer gets it right with the shaders, this
should be just what we want if I understood the problem correctly.)

Added a default-small.mesh which is just the default mesh but half the size
so we can have a couple different meshes for debugging.

Narrowed the types to a collection of mesh-component utility functions.
For example, we didn't need to pass in the whole component when just
the location would do. This reduces the impact of changes to code when
altering the definition mesh component internal slots, etc.