Integrated Step 4 into Step 3! Finally!
Replaced the component-active-view slot in core-state with a
component-active-by-type-view slot.  The former was a place holder
for before this commit. The latter is a "type table". A type table is
a hash table keyed by a type (a symbol) and who value is another hash
table. The second hash table is keyed by an instance, and the value is
the same instance. The type key given to a type-table is canonicalized.
This means if the user referenced it in the component-dependency graph,
it is unchanged, but if not, it is converted to the unknown-type-id
which is stored in the annotation for the component-dependency graph.

Upgraded EXCEUTE-FLOW to handle policies returned by the selector function.
> If the policy is :identity-policy, then the selection can be:
> a single hash table instance
> a single instance of some class
> a list of things that are either hash tables or class instances.
>
> If the policy is :type-policy, then the selection can be:
> a single type-table instance

Then, fixed execute-flow to understand how to act on a selection
when the policy is :type-policy.

Fixed the default.flow to use the :type-policy properly and had
to slightly rewrite some flow-state actions to be correct wrt this
new feature.

Fixed up the component-dependency *.gph files to hold real data now given
the component types we currently have.

Moved the initial scene tree creation concerning @universe to the
preparation of the scenes exxtension. The graphs are required to be
loaded before ANYTHING can be inserted into the scene tree, including
the universe!

Moved some code around to differnet places that seemed more appropriate.