Moderate refactor of how actors/components are spawned and initialized. This fixes a problem where if inside of an INITIALIZR-COMPONENT protocol method, you spawn an actor with some components, those actors/components would be mistakenly lost. Also, fixed some of a future problem for destroying component which requires the same behavior.
This was fixed by creating very specific states in the core-state for
actors/component that are: preinit, init, active, pre-destroy, destroy.
Added some more explcitly named functions to handle the state changes and
got rid of poorly named one that previously did it.

Then I fixed the default.flow to have a new initialization-phase
flow which systematically looped an algorithm to initialize the
actors/components, but only once, and then see if that generated any
more actor/components which must be initialized. Once the cycle hit
a fixed point of no more generated actors/components, we moved on in
the flow to the newly minted pile of active components,

Fixed SPAWN-ACTOR to reasonably allow a DWIM setting of the actor's parent.

Small and easy optimization in the initialize-phase flow to skip a large
portion of the call-flow if there are no actors/component to initialize.