Figured out that GLTF2 textures use a different coordinate system then OGL. Also noticed that the default texture wrapping is :repeat, as opposed to :clamp-to-edge.
So for now, I hacked the texture loader to NOT flip-y the loaded texture
and to default to :repeat for the wrapping. I'll need to get this data from
the glb file and package it up somehow.

I also need to think about this from other perspectives about mixing and
matching different formats like glb, or just using the system in a raw way.